Move screenrecorder to ui
